Added service implementations
